refactor(LineAcceptanceTest): 지하철 노선 생성 시 생성된 지하철역 아이디를 참조하도록 개선

+++ src/test/java/nextstep/subway/line/LineAcceptanceTestFixture.java
@@ -11,7 +11,7 @@ import org.springframework.http.MediaType;
 public class LineAcceptanceTestFixture {
     public static final String LINE_API_MAIN_PATH = "/lines";
 
-    public static ExtractableResponse<Response> createLine(String name, String color, int distance, int upStationId, int downStationId) {
+    public static ExtractableResponse<Response> createLine(String name, String color, int distance, Long upStationId, Long downStationId) {
         Map<String, Object> params = new HashMap<>();
         params.put("name", name);
         params.put("color", color);
